refactor: inline server helpers
