--- packages/node-fetch-server/src/lib/promise.ts
@@ -1,5 +1,0 @@
-export function isPromiseLike<value>(
-  value: value | PromiseLike<value>,
-): value is PromiseLike<value> {
-  return typeof (value as { then?: unknown }).then === 'function'
-}
