+++ packages/node-fetch-server/src/lib/request-listener.ts
@@ -2,9 +2,12 @@ import type * as http from 'node:http'
 import type * as http2 from 'node:http2'
 
 import type { ClientAddress, ErrorHandler, FetchHandler } from './fetch-handler.ts'
-import { createErrorResponse, defaultErrorHandler } from './error-response.ts'
 import { createLazyRequestFactory } from './lazy-request.ts'
-import { isPromiseLike } from './promise.ts'
+
+// "Internal Server Error"
+const internalServerErrorBody = [
+  73, 110, 116, 101, 114, 110, 97, 108, 32, 83, 101, 114, 118, 101, 114, 32, 69, 114, 114, 111, 114,
+]
 
 /**
  * Options for creating a Node.js request listener.
@@ -141,6 +144,33 @@ async function sendErrorResponse(
 ): Promise<void> {
   let response = await createErrorResponse(onError, error)
   await sendResponse(res, response)
+}
+
+async function createErrorResponse(onError: ErrorHandler, error: unknown): Promise<Response> {
+  try {
+    return (await onError(error)) ?? internalServerError()
+  } catch (error) {
+    console.error(`There was an error in the error handler: ${error}`)
+    return internalServerError()
+  }
+}
+
+function defaultErrorHandler(error: unknown): Response {
+  console.error(error)
+  return internalServerError()
+}
+
+function internalServerError(): Response {
+  return new Response(new Uint8Array(internalServerErrorBody), {
+    status: 500,
+    headers: {
+      'Content-Type': 'text/plain',
+    },
+  })
+}
+
+function isPromiseLike<value>(value: value | PromiseLike<value>): value is PromiseLike<value> {
+  return typeof (value as { then?: unknown }).then === 'function'
 }
 
 /**
