+++ packages/node-serve/src/lib/server.ts
@@ -1,10 +1,13 @@
 import { STATUS_CODES } from 'node:http'
 import * as uWS from 'uWebSockets.js'
 
-import { createErrorResponse, defaultErrorHandler } from './error-response.ts'
 import type { ClientAddress, ErrorHandler, FetchHandler } from './fetch-handler.ts'
-import { isPromiseLike } from './promise.ts'
 import { createUwsRequest, type UwsRequestOptions, type UwsResponseState } from './uws-request.ts'
+
+// "Internal Server Error"
+const internalServerErrorBody = [
+  73, 110, 116, 101, 114, 110, 97, 108, 32, 83, 101, 114, 118, 101, 114, 32, 69, 114, 114, 111, 114,
+]
 
 export interface UwsRequestHandlerOptions extends UwsRequestOptions {
   /**
@@ -324,6 +327,33 @@ async function sendErrorResponse(
 ): Promise<void> {
   let response = await createErrorResponse(onError, error)
   await sendUwsResponse(res, state, method, response)
+}
+
+async function createErrorResponse(onError: ErrorHandler, error: unknown): Promise<Response> {
+  try {
+    return (await onError(error)) ?? internalServerError()
+  } catch (error) {
+    console.error(`There was an error in the error handler: ${error}`)
+    return internalServerError()
+  }
+}
+
+function defaultErrorHandler(error: unknown): Response {
+  console.error(error)
+  return internalServerError()
+}
+
+function internalServerError(): Response {
+  return new Response(new Uint8Array(internalServerErrorBody), {
+    status: 500,
+    headers: {
+      'Content-Type': 'text/plain',
+    },
+  })
+}
+
+function isPromiseLike<value>(value: value | PromiseLike<value>): value is PromiseLike<value> {
+  return typeof (value as { then?: unknown }).then === 'function'
 }
 
 function createUwsResponseState(res: uWS.HttpResponse): UwsResponseState {
