+++ packages/node-serve/src/lib/uws-request.ts
@@ -1,11 +1,5 @@
 import type { HttpRequest, HttpResponse } from 'uWebSockets.js'
 
-import {
-  bodyUnusable,
-  bufferToArrayBuffer,
-  bufferToBytes,
-  requestMethodCanHaveBody,
-} from './request-body.ts'
 import { createUwsHeaders } from './uws-headers.ts'
 
 export interface UwsRequestOptions {
@@ -211,6 +205,24 @@ class UwsRequest implements Request {
 }
 
 Object.setPrototypeOf(UwsRequest.prototype, Request.prototype)
+
+function requestMethodCanHaveBody(method: string): boolean {
+  return method !== 'GET' && method !== 'HEAD'
+}
+
+function bufferToArrayBuffer(buffer: Buffer): ArrayBuffer {
+  return bufferToBytes(buffer).buffer
+}
+
+function bufferToBytes(buffer: Buffer): Uint8Array<ArrayBuffer> {
+  let bytes = new Uint8Array(buffer.byteLength)
+  bytes.set(buffer)
+  return bytes
+}
+
+function bodyUnusable(): TypeError {
+  return new TypeError('Body is unusable: Body has already been read')
+}
 
 function readUwsRequestBody(res: HttpResponse, state: UwsResponseState): Promise<Buffer> {
   return new Promise((resolve, reject) => {
